HSEARCH-3787 Use always long type to store timeout millis
Different commit, not sure it the right choice

+++ backend/elasticsearch/src/main/java/org/hibernate/search/backend/elasticsearch/search/query/impl/Elasticsearch7SearchResultExtractor.java
@@ -81,8 +81,8 @@ class Elasticsearch7SearchResultExtractor<H> implements ElasticsearchSearchResul
 		Integer took = TOOK_ACCESSOR.get( responseBody ).get();
 		Boolean timedOut = TIMED_OUT_ACCESSOR.get( responseBody ).get();
 
-		Integer remainingTimeToHardTimeout = ( hardTimeoutInMilliseconds == null ) ? null :
-				Math.toIntExact( hardTimeoutInMilliseconds ) - took;
+		Long remainingTimeToHardTimeout = ( hardTimeoutInMilliseconds == null ) ? null :
+				hardTimeoutInMilliseconds - took;
 
 		return new ElasticsearchLoadableSearchResult<>(
 				extractContext,
